backend/requirements.txt
# Framework Web
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Base de Datos
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Machine Learning
tensorflow==2.18.0
lightfm-next==1.18.7
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# CORS y Middleware
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
